Add `enable` option
An `enable` option is added as a new third argument, which defaults to
`true`. It can be set to `false` to disable geolocation. The idea is
that the option can be flipped to `true` once the user has indicated
that they want to be geolocated.

This means the hook can be made conditional without violating the rules
of hooks and without a conditional nested component which would need to
pass its data back up the tree.

Closes #51